+++ application/api/src/main/kotlin/io/raemian/api/goal/controller/response/GoalsResponse.kt
@@ -1,17 +1,8 @@
 package io.raemian.api.goal.controller.response
 
+import io.raemian.api.support.format
 import io.raemian.storage.db.core.goal.Goal
 import io.raemian.storage.db.core.sticker.StickerImage
-import java.time.LocalDate
-
-fun LocalDate.format(): String {
-    var month = this.monthValue.toString()
-    if (month.length == 1) {
-        month = "0$month"
-    }
-    
-    return "${this.year}.$month"
-}
 
 data class GoalsResponse(
     val goals: Goals,
