+++ application/api/src/main/kotlin/io/raemian/api/support/RaemianLocalDate.kt
@@ -4,6 +4,15 @@ import java.time.LocalDate
 import java.time.Month
 import java.time.Year
 
+fun LocalDate.format(): String {
+    var month = this.monthValue.toString()
+    if (month.length == 1) {
+        month = "0$month"
+    }
+
+    return "${this.year}.$month"
+}
+
 object RaemianLocalDate {
 
     private const val DAY_OF_MONTH = 1
